Remove unused `telemetry` and `storage` subsystems
