+++ packages/sdk/src/index.ts
@@ -6,7 +6,7 @@
 export * from './adapters';
 export * from './animation';
 export * from './math';
-export * from './services';
+export * from './connection';
 export * from './webHost';
 export * from './types/runtime';
 export * from './types/network';
