--- packages/sdk/src/services/connection/index.ts
@@ -1,10 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export * from './connection';
-export * from './nullConnection';
-export * from './eventedConnection';
-export * from './webSocket';
-export * from './pipe';
