--- packages/sdk/src/services/index.ts
@@ -1,8 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export * from './storage';
-export * from './connection';
-export * from './telemetry';
