--- packages/sdk/src/services/storage/fileStorage.ts
@@ -1,6 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-// TODO: Make a file storage
