--- packages/sdk/src/services/storage/index.ts
@@ -1,12 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export * from './storage';
-export * from './nullStorage';
-// export * from './fileStorage';
-// export * from './fileStorageProvider';
-export * from './memoryStorage';
-export * from './memoryStorageProvider';
-export * from './storageProvider';
